add p2p.js
